Consolidate API utilities into unified api-helpers module
Merge route-helpers.ts and api/responses.ts into single api-helpers.ts file:

## Module Consolidation
- Create src/lib/api-helpers.ts with combined functionality
- Remove src/lib/route-helpers.ts and src/lib/api/ directory
- Update all imports across 40+ files to use consolidated module

## Unified API Utilities
- Request handling: withParams(), withErrorHandling(), parameter extraction
- Response formatting: createSuccessResponse(), error response helpers
- Error processing: Enhanced createInternalError() with HttpError support
- Content-type handling: JSON, YAML, binary content processing

## Benefits
- Single import: All API utilities from one module
- Logical grouping: Request/response lifecycle in same file
- Reduced complexity: Eliminates scattered API utility files
- Better maintainability: Related functionality co-located

## Import Simplification
Before:
```typescript
import { withParams } from '@src/lib/route-helpers.js';
import { createSuccessResponse } from '@src/lib/api/responses.js';
```

After:
```typescript
import { withParams, createSuccessResponse } from '@src/lib/api-helpers.js';
```

Complete consolidation of API request/response utilities with maintained functionality.
