+++ src/lib/observers/route-integration.ts
@@ -8,7 +8,7 @@ import type { Context } from 'hono';
 import type { System } from '@src/lib/system.js';
 import type { OperationType, ObserverResult } from '@src/lib/observers/types.js';
 import { ObserverRunner } from '@src/lib/observers/runner.js';
-import { createSuccessResponse, createValidationError } from '@src/lib/api/responses.js';
+import { createSuccessResponse, createValidationError } from '@src/lib/api-helpers.js';
 
 /**
  * Execute observer pipeline for a single record operation
